requirements.txt
google-generativeai
youtube_transcript_api
pytube
pathlib
textwrap
IPython
langchain-google-vertexai
python-dotenv
